feat: notification bell, session expiry handling, SEO metadata, and optimistic React Query updates
Closes #1516, #1517, #1518, #1519
