+++ frontend/app/contact/ContactClient.tsx
@@ -1,0 +1,504 @@
+"use client";
+
+import React, { useState } from "react";
+import {
+  Building2,
+  Mail,
+  Phone,
+  MapPin,
+  Send,
+  CheckCircle,
+  AlertCircle,
+  Clock,
+  MessageSquare,
+  User,
+  Briefcase,
+  Globe,
+  Twitter,
+  Linkedin,
+  Github,
+  Link,
+} from "lucide-react";
+import { apiClient } from "@/lib/apiClient";
+import { toast } from "sonner";
+
+const ContactUsClient = () => {
+  const [formData, setFormData] = useState({
+    fullName: "",
+    email: "",
+    phone: "",
+    company: "",
+    subject: "",
+    message: "",
+  });
+  const [errors, setErrors] = useState<Record<string, string>>({});
+  const [isSubmitted, setIsSubmitted] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
+
+  const handleInputChange = (field: string, value: string) => {
+    setFormData((prev) => ({ ...prev, [field]: value }));
+    if (errors[field]) {
+      setErrors((prev) => {
+        const newErrors = { ...prev };
+        delete newErrors[field];
+        return newErrors;
+      });
+    }
+  };
+
+  const validateForm = () => {
+    const newErrors: Record<string, string> = {};
+
+    if (!formData.fullName.trim()) newErrors.fullName = "Full name is required";
+    if (!formData.email.trim()) newErrors.email = "Email is required";
+    else if (!/\S+@\S+\.\S+/.test(formData.email))
+      newErrors.email = "Invalid email format";
+    if (!formData.subject.trim()) newErrors.subject = "Subject is required";
+    if (!formData.message.trim()) newErrors.message = "Message is required";
+    else if (formData.message.trim().length < 10)
+      newErrors.message = "Message must be at least 10 characters";
+
+    setErrors(newErrors);
+    return Object.keys(newErrors).length === 0;
+  };
+
+  const handleSubmit = async () => {
+    if (!validateForm()) return;
+
+    setIsLoading(true);
+    try {
+      await apiClient.post("/contact", {
+        fullName: formData.fullName,
+        email: formData.email,
+        phone: formData.phone || undefined,
+        company: formData.company || undefined,
+        subject: formData.subject,
+        message: formData.message,
+      });
+      setIsSubmitted(true);
+    } catch (error) {
+      const msg =
+        error instanceof Error ? error.message : "Failed to send message";
+      toast.error(msg);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  const contactInfo = [
+    {
+      icon: <Mail className="h-6 w-6" />,
+      title: "Email",
+      content: "support@managehub.com",
+      link: "mailto:support@managehub.com",
+    },
+    {
+      icon: <Phone className="h-6 w-6" />,
+      title: "Phone",
+      content: "+234 800 000 0000",
+      link: "tel:+2348000000000",
+    },
+    {
+      icon: <MapPin className="h-6 w-6" />,
+      title: "Office",
+      content: "Abuja, FCT, Nigeria",
+      link: null,
+    },
+    {
+      icon: <Clock className="h-6 w-6" />,
+      title: "Business Hours",
+      content: "Mon - Fri: 9:00 AM - 6:00 PM WAT",
+      link: null,
+    },
+  ];
+
+  const socialLinks = [
+    { icon: <Twitter className="h-5 w-5" />, label: "Twitter", href: "#" },
+    { icon: <Linkedin className="h-5 w-5" />, label: "LinkedIn", href: "#" },
+    { icon: <Github className="h-5 w-5" />, label: "GitHub", href: "#" },
+  ];
+
+  if (isSubmitted) {
+    return (
+      <div className="min-h-screen bg-[#faf9f7] flex items-center justify-center px-4 py-12">
+        <div className="max-w-md w-full space-y-8">
+          <div className="text-center">
+            <div className="flex justify-center items-center mb-6">
+              <div className="bg-gray-900 p-3 rounded-xl">
+                <Building2 className="h-8 w-8 text-white" />
+              </div>
+            </div>
+            <h1 className="text-3xl font-bold text-gray-900">ManageHub</h1>
+          </div>
+
+          <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-8">
+            <div className="text-center space-y-6">
+              <div className="flex justify-center">
+                <div className="bg-green-100 p-4 rounded-full">
+                  <CheckCircle className="h-12 w-12 text-green-600" />
+                </div>
+              </div>
+
+              <div className="space-y-2">
+                <h2 className="text-2xl font-bold text-gray-900">
+                  Message Sent Successfully!
+                </h2>
+                <p className="text-gray-600">
+                  Thank you for reaching out to us. We&apos;ve received your message
+                  and our team will get back to you within 24-48 hours.
+                </p>
+              </div>
+
+              <div className="bg-gray-50 border border-gray-200 rounded-lg p-4">
+                <p className="text-sm text-gray-900">
+                  We&apos;ve sent a confirmation email to{" "}
+                  <span className="font-medium">{formData.email}</span>
+                </p>
+              </div>
+
+              <button
+                onClick={() => {
+                  setIsSubmitted(false);
+                  setFormData({
+                    fullName: "",
+                    email: "",
+                    phone: "",
+                    company: "",
+                    subject: "",
+                    message: "",
+                  });
+                }}
+                className="w-full bg-gray-900 text-white py-3 px-4 rounded-lg font-medium hover:bg-gray-800 transition-colors"
+              >
+                Send Another Message
+              </button>
+
+              <Link
+                href="/"
+                className="block text-gray-700 hover:text-gray-900 font-medium transition-colors"
+              >
+                Back to Home
+              </Link>
+            </div>
+          </div>
+        </div>
+      </div>
+    );
+  }
+
+  return (
+    <div className="min-h-screen bg-[#faf9f7] py-12 px-4 sm:px-6 lg:px-8">
+      <div className="max-w-7xl mx-auto">
+        <div className="text-center mb-12">
+          <div className="flex justify-center items-center mb-6">
+            <div className="bg-gray-900 p-3 rounded-xl">
+              <Building2 className="h-8 w-8 text-white" />
+            </div>
+          </div>
+          <h1 className="text-4xl font-bold text-gray-900 mb-4">
+            Get in Touch
+          </h1>
+          <p className="text-xl text-gray-600 max-w-2xl mx-auto">
+            Have questions about ManageHub? We&apos;d love to hear from you. Send us
+            a message and we&apos;ll respond as soon as possible.
+          </p>
+        </div>
+
+        <div className="grid lg:grid-cols-3 gap-8">
+          <div className="lg:col-span-2">
+            <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-8">
+              <div className="flex items-center mb-6">
+                <MessageSquare className="h-6 w-6 text-gray-700 mr-2" />
+                <h2 className="text-2xl font-bold text-gray-900">
+                  Send us a Message
+                </h2>
+              </div>
+
+              <div className="space-y-6">
+                <div>
+                  <label
+                    htmlFor="fullName"
+                    className="block text-sm font-medium text-gray-700 mb-2"
+                  >
+                    Full Name <span className="text-red-500">*</span>
+                  </label>
+                  <div className="relative">
+                    <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                      <User className="h-5 w-5 text-gray-400" />
+                    </div>
+                    <input
+                      id="fullName"
+                      type="text"
+                      value={formData.fullName}
+                      onChange={(e) =>
+                        handleInputChange("fullName", e.target.value)
+                      }
+                      className={`block w-full pl-10 pr-3 py-3 border rounded-lg placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-gray-300 focus:border-transparent transition-all ${
+                        errors.fullName ? "border-red-500" : "border-gray-300"
+                      }`}
+                      placeholder="Enter your full name"
+                    />
+                  </div>
+                  {errors.fullName && (
+                    <p className="mt-1 text-sm text-red-600 flex items-center">
+                      <AlertCircle className="h-4 w-4 mr-1" />
+                      {errors.fullName}
+                    </p>
+                  )}
+                </div>
+
+                <div className="grid md:grid-cols-2 gap-6">
+                  <div>
+                    <label
+                      htmlFor="email"
+                      className="block text-sm font-medium text-gray-700 mb-2"
+                    >
+                      Email Address <span className="text-red-500">*</span>
+                    </label>
+                    <div className="relative">
+                      <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                        <Mail className="h-5 w-5 text-gray-400" />
+                      </div>
+                      <input
+                        id="email"
+                        type="email"
+                        value={formData.email}
+                        onChange={(e) =>
+                          handleInputChange("email", e.target.value)
+                        }
+                        className={`block w-full pl-10 pr-3 py-3 border rounded-lg placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-gray-300 focus:border-transparent transition-all ${
+                          errors.email ? "border-red-500" : "border-gray-300"
+                        }`}
+                        placeholder="your.email@example.com"
+                      />
+                    </div>
+                    {errors.email && (
+                      <p className="mt-1 text-sm text-red-600 flex items-center">
+                        <AlertCircle className="h-4 w-4 mr-1" />
+                        {errors.email}
+                      </p>
+                    )}
+                  </div>
+
+                  <div>
+                    <label
+                      htmlFor="phone"
+                      className="block text-sm font-medium text-gray-700 mb-2"
+                    >
+                      Phone Number{" "}
+                      <span className="text-gray-400 text-xs">(Optional)</span>
+                    </label>
+                    <div className="relative">
+                      <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                        <Phone className="h-5 w-5 text-gray-400" />
+                      </div>
+                      <input
+                        id="phone"
+                        type="tel"
+                        value={formData.phone}
+                        onChange={(e) =>
+                          handleInputChange("phone", e.target.value)
+                        }
+                        className="block w-full pl-10 pr-3 py-3 border border-gray-300 rounded-lg placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-gray-300 focus:border-transparent transition-all"
+                        placeholder="+234 800 000 0000"
+                      />
+                    </div>
+                  </div>
+                </div>
+
+                <div>
+                  <label
+                    htmlFor="company"
+                    className="block text-sm font-medium text-gray-700 mb-2"
+                  >
+                    Company/Organization{" "}
+                    <span className="text-gray-400 text-xs">(Optional)</span>
+                  </label>
+                  <div className="relative">
+                    <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                      <Briefcase className="h-5 w-5 text-gray-400" />
+                    </div>
+                    <input
+                      id="company"
+                      type="text"
+                      value={formData.company}
+                      onChange={(e) =>
+                        handleInputChange("company", e.target.value)
+                      }
+                      className="block w-full pl-10 pr-3 py-3 border border-gray-300 rounded-lg placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-gray-300 focus:border-transparent transition-all"
+                      placeholder="Your company name"
+                    />
+                  </div>
+                </div>
+
+                <div>
+                  <label
+                    htmlFor="subject"
+                    className="block text-sm font-medium text-gray-700 mb-2"
+                  >
+                    Subject <span className="text-red-500">*</span>
+                  </label>
+                  <div className="relative">
+                    <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                      <Globe className="h-5 w-5 text-gray-400" />
+                    </div>
+                    <input
+                      id="subject"
+                      type="text"
+                      value={formData.subject}
+                      onChange={(e) =>
+                        handleInputChange("subject", e.target.value)
+                      }
+                      className={`block w-full pl-10 pr-3 py-3 border rounded-lg placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-gray-300 focus:border-transparent transition-all ${
+                        errors.subject ? "border-red-500" : "border-gray-300"
+                      }`}
+                      placeholder="How can we help you?"
+                    />
+                  </div>
+                  {errors.subject && (
+                    <p className="mt-1 text-sm text-red-600 flex items-center">
+                      <AlertCircle className="h-4 w-4 mr-1" />
+                      {errors.subject}
+                    </p>
+                  )}
+                </div>
+
+                <div>
+                  <label
+                    htmlFor="message"
+                    className="block text-sm font-medium text-gray-700 mb-2"
+                  >
+                    Message <span className="text-red-500">*</span>
+                  </label>
+                  <textarea
+                    id="message"
+                    rows={6}
+                    value={formData.message}
+                    onChange={(e) =>
+                      handleInputChange("message", e.target.value)
+                    }
+                    className={`block w-full px-3 py-3 border rounded-lg placeholder-gray-400 focus:outline-none focus:ring-2 focus:ring-gray-300 focus:border-transparent transition-all resize-none ${
+                      errors.message ? "border-red-500" : "border-gray-300"
+                    }`}
+                    placeholder="Tell us more about your inquiry..."
+                  ></textarea>
+                  {errors.message && (
+                    <p className="mt-1 text-sm text-red-600 flex items-center">
+                      <AlertCircle className="h-4 w-4 mr-1" />
+                      {errors.message}
+                    </p>
+                  )}
+                  <p className="mt-2 text-sm text-gray-500">
+                    {formData.message.length} characters
+                  </p>
+                </div>
+
+                <button
+                  onClick={handleSubmit}
+                  disabled={isLoading}
+                  className="w-full bg-gray-900 text-white py-3 px-4 rounded-lg font-medium hover:bg-gray-800 focus:outline-none focus:ring-2 focus:ring-gray-300 focus:ring-offset-2 transition-colors disabled:opacity-50 disabled:cursor-not-allowed flex items-center justify-center"
+                >
+                  {isLoading ? (
+                    <>
+                      <div className="animate-spin rounded-full h-5 w-5 border-b-2 border-white mr-2"></div>
+                      Sending...
+                    </>
+                  ) : (
+                    <>
+                      <Send className="h-5 w-5 mr-2" />
+                      Send Message
+                    </>
+                  )}
+                </button>
+              </div>
+            </div>
+          </div>
+
+          <div className="space-y-6">
+            <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-6">
+              <h3 className="text-lg font-bold text-gray-900 mb-4">
+                Contact Information
+              </h3>
+              <div className="space-y-4">
+                {contactInfo.map((info, index) => (
+                  <div key={index} className="flex items-start">
+                    <div className="bg-gray-100 p-2 rounded-lg flex-shrink-0">
+                      <div className="text-gray-700">{info.icon}</div>
+                    </div>
+                    <div className="ml-4">
+                      <p className="text-sm font-medium text-gray-900">
+                        {info.title}
+                      </p>
+                      {info.link ? (
+                        <a
+                          href={info.link}
+                          className="text-sm text-gray-600 hover:text-gray-700 transition-colors"
+                        >
+                          {info.content}
+                        </a>
+                      ) : (
+                        <p className="text-sm text-gray-600">{info.content}</p>
+                      )}
+                    </div>
+                  </div>
+                ))}
+              </div>
+            </div>
+
+            <div className="bg-white rounded-xl shadow-sm border border-gray-200 p-6">
+              <h3 className="text-lg font-bold text-gray-900 mb-4">
+                Follow Us
+              </h3>
+              <div className="flex gap-3">
+                {socialLinks.map((social, index) => (
+                  <a
+                    key={index}
+                    href={social.href}
+                    className="bg-gray-100 p-3 rounded-lg hover:bg-gray-900 hover:text-white text-gray-600 transition-colors"
+                    aria-label={social.label}
+                  >
+                    {social.icon}
+                  </a>
+                ))}
+              </div>
+            </div>
+
+            <div className="bg-gradient-to-br from-gray-50 to-gray-100 border border-gray-200 rounded-xl p-6">
+              <h3 className="text-lg font-bold text-gray-900 mb-2">
+                Need Quick Help?
+              </h3>
+              <p className="text-sm text-gray-700 mb-4">
+                Check out our documentation and FAQ section for instant answers
+                to common questions.
+              </p>
+              <a
+                href="#"
+                className="inline-flex items-center text-sm font-medium text-gray-700 hover:text-gray-600"
+              >
+                Visit Help Center
+                <span className="ml-1">→</span>
+              </a>
+            </div>
+          </div>
+        </div>
+
+        <div className="text-center mt-12 text-sm text-gray-500">
+          <p>© 2026 ManageHub. All rights reserved.</p>
+          <div className="mt-2 space-x-4">
+            <a href="#" className="hover:text-gray-700">
+              Privacy Policy
+            </a>
+            <a href="#" className="hover:text-gray-700">
+              Terms of Service
+            </a>
+            <a href="#" className="hover:text-gray-700">
+              Support
+            </a>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default ContactUsClient;
