+++ frontend/lib/apiClient.ts
@@ -3,9 +3,9 @@ const API_BASE_URL =
 
 export class ApiError extends Error {
   status?: number;
-  data?: any;
+  data?: unknown;
 
-  constructor(message: string, status?: number, data?: any) {
+  constructor(message: string, status?: number, data?: unknown) {
     super(message);
     this.name = "ApiError";
     this.status = status;
@@ -13,9 +13,16 @@ export class ApiError extends Error {
   }
 }
 
+type RequestInterceptor = (config: RequestInit) => RequestInit | Promise<RequestInit>;
+type ResponseInterceptor = (response: Response) => Response | Promise<Response>;
+type ResponseErrorInterceptor = (error: unknown) => unknown;
+
 class ApiClient {
   private baseURL: string;
   private token: string | null = null;
+  private requestInterceptors: RequestInterceptor[] = [];
+  private responseInterceptors: ResponseInterceptor[] = [];
+  private responseErrorInterceptors: ResponseErrorInterceptor[] = [];
 
   constructor(baseURL: string) {
     this.baseURL = baseURL;
@@ -25,41 +32,63 @@ class ApiClient {
     this.token = token;
   }
 
+  addRequestInterceptor(interceptor: RequestInterceptor) {
+    this.requestInterceptors.push(interceptor);
+  }
+
+  addResponseInterceptor(
+    onFulfilled: ResponseInterceptor,
+    onRejected?: ResponseErrorInterceptor
+  ) {
+    this.responseInterceptors.push(onFulfilled);
+    if (onRejected) {
+      this.responseErrorInterceptors.push(onRejected);
+    }
+  }
+
   private async request<T>(
     endpoint: string,
     options: RequestInit = {}
   ): Promise<T> {
     const url = `${this.baseURL}${endpoint}`;
 
-    const headers: Record<string, string> = {
-      "Content-Type": "application/json",
-      ...(options.headers as Record<string, string>),
+    let config: RequestInit = {
+      ...options,
+      headers: {
+        "Content-Type": "application/json",
+        ...(options.headers as Record<string, string>),
+      },
+      credentials: "include",
     };
 
     if (this.token) {
-      headers["Authorization"] = `Bearer ${this.token}`;
+      (config.headers as Record<string, string>)["Authorization"] =
+        `Bearer ${this.token}`;
     }
 
-    const config: RequestInit = {
-      ...options,
-      headers,
-    };
+    for (const interceptor of this.requestInterceptors) {
+      config = await interceptor(config);
+    }
 
     try {
-      const response = await fetch(url, config);
+      let response = await fetch(url, config);
+
+      for (const interceptor of this.responseInterceptors) {
+        response = await interceptor(response);
+      }
 
       if (!response.ok) {
         if (response.status === 401 && typeof window !== "undefined") {
           window.dispatchEvent(new Event("session-expired"));
         }
-        let errorData: any = {};
+        let errorData: Record<string, unknown> = {};
         try {
           errorData = await response.json();
         } catch {
           errorData = {};
         }
         throw new ApiError(
-          errorData.message || "An API error occurred",
+          (errorData.message as string) || "An API error occurred",
           response.status,
           errorData
         );
@@ -67,6 +96,11 @@ class ApiClient {
 
       return await response.json();
     } catch (error) {
+      for (const interceptor of this.responseErrorInterceptors) {
+        const result = interceptor(error);
+        if (result !== undefined) return result;
+      }
+
       if (error instanceof ApiError) {
         throw error;
       }
@@ -86,7 +120,6 @@ class ApiClient {
   async post<T, D = unknown>(endpoint: string, data?: D): Promise<T> {
     return this.request<T>(endpoint, {
       method: "POST",
-      credentials: "include",
       body: data ? JSON.stringify(data) : undefined,
     });
   }
