+++ frontend/lib/react-query/hooks/bookings/useCancelBooking.ts
@@ -14,12 +14,41 @@ export const useCancelBooking = () => {
       apiClient.patch<{ success: boolean; data: Booking }>(
         `/bookings/${bookingId}/cancel`
       ),
+    onMutate: async (bookingId) => {
+      await queryClient.cancelQueries({
+        queryKey: queryKeys.bookings.all,
+      });
+
+      const previousBookings = queryClient.getQueriesData({
+        queryKey: queryKeys.bookings.all,
+      });
+
+      queryClient.setQueriesData<{ data: Booking[]; meta: { total: number; page: number; limit: number; totalPages: number } }>(
+        { queryKey: queryKeys.bookings.mine({}) },
+        (old) => {
+          if (!old) return old;
+          return {
+            ...old,
+            data: old.data.map((b) =>
+              b.id === bookingId ? { ...b, status: "CANCELLED" as const } : b
+            ),
+          };
+        }
+      );
+
+      return { previousBookings };
+    },
+    onError: (_err, _bookingId, context) => {
+      if (context?.previousBookings) {
+        for (const [key, data] of context.previousBookings) {
+          queryClient.setQueryData(key, data);
+        }
+      }
+      toast.error("Failed to cancel booking");
+    },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: queryKeys.bookings.all });
       toast.success("Booking cancelled");
-    },
-    onError: (error: Error) => {
-      toast.error(error.message || "Failed to cancel booking");
     },
   });
 };
