+++ frontend/lib/react-query/hooks/membership/useMembershipPlans.ts
@@ -1,8 +1,9 @@
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
+import { useQuery, useMutation, useQueryClient } from "@tanstack/react-query";
+import { toast } from "sonner";
 
 // --- TYPES ---
 
-export type BillingCycle = 'MONTHLY' | 'QUARTERLY' | 'YEARLY';
+export type BillingCycle = "MONTHLY" | "QUARTERLY" | "YEARLY";
 
 export interface MembershipPlan {
   id: string;
@@ -15,7 +16,7 @@ export interface MembershipPlan {
   guestPassesPerMonth: number;
   displayOrder: number;
   isActive: boolean;
-  activeSubscribersCount?: number; // Added to fix the TypeScript error
+  activeSubscribersCount?: number;
   createdAt?: string;
   updatedAt?: string;
 }
@@ -35,14 +36,16 @@ export interface PlanFormPayload {
 
 // --- QUERY & MUTATION HOOKS ---
 
-export const MEMBERSHIP_PLANS_QUERY_KEY = ['membership-plans'];
+export const MEMBERSHIP_PLANS_QUERY_KEY = ["membership-plans"];
 
 export function useGetMembershipPlans(includeInactive: boolean = false) {
   return useQuery<MembershipPlan[]>({
     queryKey: [...MEMBERSHIP_PLANS_QUERY_KEY, { includeInactive }],
     queryFn: async () => {
-      const res = await fetch(`/api/admin/membership-plans?includeInactive=${includeInactive}`);
-      if (!res.ok) throw new Error('Failed to fetch membership plans');
+      const res = await fetch(
+        `/api/admin/membership-plans?includeInactive=${includeInactive}`
+      );
+      if (!res.ok) throw new Error("Failed to fetch membership plans");
       return res.json();
     },
   });
@@ -53,16 +56,56 @@ export function useCreatePlan() {
 
   return useMutation({
     mutationFn: async (payload: PlanFormPayload) => {
-      const res = await fetch('/api/admin/membership-plans', {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
+      const res = await fetch("/api/admin/membership-plans", {
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
         body: JSON.stringify(payload),
       });
-      if (!res.ok) throw new Error('Failed to create membership plan');
+      if (!res.ok) throw new Error("Failed to create membership plan");
       return res.json();
+    },
+    onMutate: async (payload) => {
+      await queryClient.cancelQueries({
+        queryKey: MEMBERSHIP_PLANS_QUERY_KEY,
+      });
+
+      const previousPlans = queryClient.getQueriesData<MembershipPlan[]>({
+        queryKey: MEMBERSHIP_PLANS_QUERY_KEY,
+      });
+
+      queryClient.setQueriesData<MembershipPlan[]>(
+        { queryKey: MEMBERSHIP_PLANS_QUERY_KEY },
+        (old) => {
+          if (!old) return old;
+          const optimisticPlan: MembershipPlan = {
+            id: `temp-${Date.now()}`,
+            name: payload.name || "",
+            description: payload.description || "",
+            priceKobo: payload.priceKobo || 0,
+            billingCycle: payload.billingCycle || "MONTHLY",
+            features: payload.features || [],
+            bookingHoursIncluded: payload.bookingHoursIncluded || 0,
+            guestPassesPerMonth: payload.guestPassesPerMonth || 0,
+            displayOrder: payload.displayOrder || 0,
+            isActive: payload.isActive ?? true,
+          };
+          return [...old, optimisticPlan];
+        }
+      );
+
+      return { previousPlans };
+    },
+    onError: (_err, _payload, context) => {
+      if (context?.previousPlans) {
+        for (const [key, data] of context.previousPlans) {
+          queryClient.setQueryData(key, data);
+        }
+      }
+      toast.error("Failed to create membership plan");
     },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: MEMBERSHIP_PLANS_QUERY_KEY });
+      toast.success("Plan created successfully");
     },
   });
 }
@@ -71,17 +114,53 @@ export function useUpdatePlan() {
   const queryClient = useQueryClient();
 
   return useMutation({
-    mutationFn: async ({ id, payload }: { id: string; payload: PlanFormPayload }) => {
+    mutationFn: async ({
+      id,
+      payload,
+    }: {
+      id: string;
+      payload: PlanFormPayload;
+    }) => {
       const res = await fetch(`/api/admin/membership-plans/${id}`, {
-        method: 'PATCH',
-        headers: { 'Content-Type': 'application/json' },
+        method: "PATCH",
+        headers: { "Content-Type": "application/json" },
         body: JSON.stringify(payload),
       });
-      if (!res.ok) throw new Error('Failed to update membership plan');
+      if (!res.ok) throw new Error("Failed to update membership plan");
       return res.json();
+    },
+    onMutate: async ({ id, payload }) => {
+      await queryClient.cancelQueries({
+        queryKey: MEMBERSHIP_PLANS_QUERY_KEY,
+      });
+
+      const previousPlans = queryClient.getQueriesData<MembershipPlan[]>({
+        queryKey: MEMBERSHIP_PLANS_QUERY_KEY,
+      });
+
+      queryClient.setQueriesData<MembershipPlan[]>(
+        { queryKey: MEMBERSHIP_PLANS_QUERY_KEY },
+        (old) => {
+          if (!old) return old;
+          return old.map((plan) =>
+            plan.id === id ? { ...plan, ...payload } : plan
+          );
+        }
+      );
+
+      return { previousPlans };
+    },
+    onError: (_err, _data, context) => {
+      if (context?.previousPlans) {
+        for (const [key, data] of context.previousPlans) {
+          queryClient.setQueryData(key, data);
+        }
+      }
+      toast.error("Failed to update membership plan");
     },
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: MEMBERSHIP_PLANS_QUERY_KEY });
+      toast.success("Plan updated successfully");
     },
   });
 }
@@ -92,10 +171,10 @@ export function useSubscribeToPlan() {
   return useMutation({
     mutationFn: async (planId: string) => {
       const res = await fetch(`/api/membership-plans/${planId}/subscribe`, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
+        method: "POST",
+        headers: { "Content-Type": "application/json" },
       });
-      if (!res.ok) throw new Error('Failed to subscribe to plan');
+      if (!res.ok) throw new Error("Failed to subscribe to plan");
       return res.json();
     },
     onSuccess: () => {
