+++ frontend/lib/react-query/hooks/notifications/useMarkAllRead.ts
@@ -3,7 +3,20 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { apiClient } from "@/lib/apiClient";
 import { queryKeys } from "@/lib/react-query/keys/queryKeys";
+import { Notification } from "@/lib/types/notification";
 import { toast } from "sonner";
+
+interface NotificationsResponse {
+  success: boolean;
+  data: Notification[];
+  meta: {
+    total: number;
+    page: number;
+    limit: number;
+    totalPages: number;
+    unreadCount: number;
+  };
+}
 
 export const useMarkAllRead = () => {
   const queryClient = useQueryClient();
@@ -11,11 +24,47 @@ export const useMarkAllRead = () => {
   return useMutation({
     mutationFn: () =>
       apiClient.patch<{ success: boolean }>("/notifications/read-all"),
+    onMutate: async () => {
+      await queryClient.cancelQueries({
+        queryKey: queryKeys.notifications.all,
+      });
+
+      const previousData = queryClient.getQueriesData<NotificationsResponse>({
+        queryKey: queryKeys.notifications.all,
+      });
+
+      queryClient.setQueriesData<NotificationsResponse>(
+        { queryKey: queryKeys.notifications.all },
+        (old) => {
+          if (!old) return old;
+          return {
+            ...old,
+            data: old.data.map((n) => ({ ...n, isRead: true })),
+            meta: {
+              ...old.meta,
+              unreadCount: 0,
+            },
+          };
+        }
+      );
+
+      return { previousData };
+    },
+    onError: (_err, _data, context) => {
+      if (context?.previousData) {
+        for (const [key, data] of context.previousData) {
+          queryClient.setQueryData(key, data);
+        }
+      }
+      toast.error("Failed to mark all as read");
+    },
     onSuccess: () => {
+      toast.success("All notifications marked as read");
+    },
+    onSettled: () => {
       queryClient.invalidateQueries({
         queryKey: queryKeys.notifications.all,
       });
-      toast.success("All notifications marked as read");
     },
   });
 };
