+++ frontend/middleware.ts
@@ -24,10 +24,12 @@ const JWT_SECRET = new TextEncoder().encode(
   process.env.JWT_SECRET || "fallback-secret"
 );
 
+const REFRESH_THRESHOLD_SECONDS = 5 * 60;
+
 async function decodeToken(token: string) {
   try {
     const { payload } = await jwtVerify(token, JWT_SECRET);
-    return payload as { sub?: string; role?: string; email?: string };
+    return payload as { sub?: string; role?: string; email?: string; exp?: number };
   } catch {
     return null;
   }
@@ -79,6 +81,25 @@ export async function middleware(request: NextRequest) {
       const response = NextResponse.redirect(loginUrl);
       response.cookies.delete("authToken");
       return response;
+    }
+
+    if (payload.exp) {
+      const nowSeconds = Math.floor(Date.now() / 1000);
+      const secondsUntilExpiry = payload.exp - nowSeconds;
+
+      if (secondsUntilExpiry <= 0) {
+        const loginUrl = new URL("/login", request.url);
+        loginUrl.searchParams.set("redirect", pathname);
+        const response = NextResponse.redirect(loginUrl);
+        response.cookies.delete("authToken");
+        return response;
+      }
+
+      if (secondsUntilExpiry <= REFRESH_THRESHOLD_SECONDS) {
+        const response = NextResponse.next();
+        response.headers.set("X-Token-Expiring-Soon", "true");
+        return response;
+      }
     }
 
     const userRole = payload.role;
